# web/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CandidateDetailOut": {
        "additionalProperties": false,
        "properties": {
          "candidate": {
            "$ref": "#/components/schemas/CandidateOut"
          },
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/EntityDecisionOut"
            },
            "title": "Decisions",
            "type": "array"
          },
          "parties": {
            "items": {
              "$ref": "#/components/schemas/PartyOut"
            },
            "title": "Parties",
            "type": "array"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "candidate",
          "run_id",
          "parties",
          "decisions"
        ],
        "title": "CandidateDetailOut",
        "type": "object"
      },
...
      "CommentIn": {
        "additionalProperties": false,
        "properties": {
          "body": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          }
        },
        "required": [
          "body"
        ],
        "title": "CommentIn",
        "type": "object"
      },
      "CommentOut": {
        "additionalProperties": false,
        "properties": {
          "author_name": {
            "title": "Author Name",
            "type": "string"
          },
          "author_user_id": {
            "format": "uuid",
            "title": "Author User Id",
            "type": "string"
          },
          "body": {
            "title": "Body",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "author_user_id",
          "author_name",
          "body",
          "created_at"
        ],
        "title": "CommentOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IssueUpdateIn": {
        "additionalProperties": false,
        "properties": {
          "clear_owner": {
            "default": false,
            "title": "Clear Owner",
            "type": "boolean"
          },
          "note": {
            "default": "",
            "maxLength": 2000,
            "title": "Note",
            "type": "string"
          },
          "owner_user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner User Id"
          },
          "status": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version"
        ],
        "title": "IssueUpdateIn",
        "type": "object"
      },
...
      "Page_IssueOut_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IssueOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "Page[IssueOut]",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Relay API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/entity-candidates/{candidate_id}": {
      "get": {
        "operationId": "get_candidate_api_v1_entity_candidates__candidate_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Candidate",
        "tags": [
          "issues"
        ]
      }
    },
...
    "/api/v1/issues/{issue_id}": {
      "get": {
        "operationId": "get_issue_api_v1_issues__issue_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "issue_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Issue Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Issue",
        "tags": [
          "workspace"
        ]
      },
      "patch": {
        "description": "Owner and workflow status. ``resolved`` is refused: only a current run resolves issues.",
        "operationId": "update_issue_api_v1_issues__issue_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "issue_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Issue Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueUpdateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Issue",
        "tags": [
          "issues"
        ]
      }
    },
    "/api/v1/issues/{issue_id}/comments": {
      "get": {
        "operationId": "list_comments_api_v1_issues__issue_id__comments_get",
        "parameters": [
          {
            "in": "path",
            "name": "issue_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Issue Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CommentOut"
                  },
                  "title": "Response List Comments Api V1 Issues  Issue Id  Comments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Comments",
        "tags": [
          "issues"
        ]
      },
      "post": {
        "operationId": "add_comment_api_v1_issues__issue_id__comments_post",
        "parameters": [
          {
            "in": "path",
            "name": "issue_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Issue Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Comment",
        "tags": [
          "issues"
        ]
      }
    },
    "/api/v1/issues/{issue_id}/history": {
      "get": {
        "operationId": "issue_history_api_v1_issues__issue_id__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "issue_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Issue Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditEventOut"
                  },
                  "title": "Response Issue History Api V1 Issues  Issue Id  History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Issue History",
        "tags": [
          "issues"
        ]
      }
    },
    "/api/v1/issues/{issue_id}/links": {
      "get": {
        "operationId": "list_links_api_v1_issues__issue_id__links_get",
        "parameters": [
          {
            "in": "path",
            "name": "issue_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Issue Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IssueLinkOut"
                  },
                  "title": "Response List Links Api V1 Issues  Issue Id  Links Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Links",
        "tags": [
          "issues"
        ]
      }
    },
...
    "/api/v1/migrations/{migration_id}/dispositions": {
      "get": {
        "operationId": "list_dispositions_api_v1_migrations__migration_id__dispositions_get",
        "parameters": [
          {
            "in": "path",
            "name": "migration_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Migration Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DispositionOut"
                  },
                  "title": "Response List Dispositions Api V1 Migrations  Migration Id  Dispositions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Dispositions",
        "tags": [
          "issues"
        ]
      }
    },
...
    "/api/v1/migrations/{migration_id}/entity-decisions": {
      "get": {
        "operationId": "list_entity_decisions_api_v1_migrations__migration_id__entity_decisions_get",
        "parameters": [
          {
            "in": "path",
            "name": "migration_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Migration Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EntityDecisionOut"
                  },
                  "title": "Response List Entity Decisions Api V1 Migrations  Migration Id  Entity Decisions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Entity Decisions",
        "tags": [
          "issues"
        ]
      }
    },
...
    "/api/v1/migrations/{migration_id}/issues": {
      "get": {
        "operationId": "list_issues_api_v1_migrations__migration_id__issues_get",
        "parameters": [
          {
            "in": "path",
            "name": "migration_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Migration Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 32,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 16,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "in": "query",
            "name": "nature",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 32,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Nature"
            }
          },
          {
            "in": "query",
            "name": "fingerprint",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fingerprint"
            }
          },
          {
            "in": "query",
            "name": "owner",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner"
            }
          },
          {
            "in": "query",
            "name": "rule",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rule"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "key",
              "enum": [
                "key",
                "amount"
              ],
              "title": "Order",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Cursor",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_IssueOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Issues",
        "tags": [
          "workspace"
        ]
      },
      "post": {
        "operationId": "create_manual_issue_api_v1_migrations__migration_id__issues_post",
        "parameters": [
          {
            "in": "path",
            "name": "migration_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Migration Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Relay-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 254,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Relay-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualIssueIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Manual Issue",
        "tags": [
          "issues"
        ]
      }
    },
...
  }
}